refactor: polish todo reminder UI and split settings into System popover
- TodoItem: date badge moved below task name to preserve text space
- TodoItem: date picker now shows as popover with Done/Cancel/Clear buttons
- Reminder scheduler: check interval 60s (was 30s) matching minute precision
- New SystemPopover: hotkey, hotkey target, open behavior, storage, version
- SettingsPopover: now only appearance (theme, opacity, colors, level, reminder)
- EditorHeader: SYS button between close and MD/TXT
