+++ src/renderer/src/editor/SettingsPopover.tsx
@@ -1,4 +1,4 @@
-import { useState, useEffect, useRef, useCallback } from 'react'
+import { useState, useEffect } from 'react'
 import { useApi } from '../hooks/use-ipc'
 
 interface SettingsPopoverProps {
@@ -12,15 +12,6 @@ export function SettingsPopover({ onClose }: SettingsPopoverProps) {
   const [fontColor, setFontColor] = useState('#1a1a1a')
   const [level, setLevel] = useState<'always' | 'normal' | 'bottom'>('always')
   const [theme, setTheme] = useState<'light' | 'dark'>('light')
-  const [shortcut, setShortcut] = useState('Alt+Space')
-  const [isRecording, setIsRecording] = useState(false)
-  const [shortcutError, setShortcutError] = useState('')
-  const [version, setVersion] = useState('')
-  const [updateStatus, setUpdateStatus] = useState<'idle' | 'checking' | 'latest' | 'available'>('idle')
-  const [latestVersion, setLatestVersion] = useState('')
-  const [storagePath, setStoragePath] = useState('')
-  const [shortcutTarget, setShortcutTarget] = useState<'command' | 'notes' | 'task'>('command')
-  const [shortcutTargetOption, setShortcutTargetOption] = useState<'last-edit' | 'new' | 'first'>('last-edit')
   const [reminderLeadTime, setReminderLeadTime] = useState(10)
 
   useEffect(() => {
@@ -31,12 +22,7 @@ export function SettingsPopover({ onClose }: SettingsPopoverProps) {
       setFontColor(ws.fontColor)
       setLevel(ws.alwaysOnTop)
       setTheme(config.theme)
-      setShortcut(config.globalShortcut || 'Alt+Space')
-      setStoragePath(config.storagePath || '')
-      setShortcutTarget(config.shortcutTarget || 'command')
-      setShortcutTargetOption(config.shortcutTargetOption || 'last-edit')
       setReminderLeadTime(config.reminderLeadTime ?? 10)
-      api.appVersion().then(setVersion)
       const { r, g, b } = hexToRgb(ws.panelColor)
       document.documentElement.style.setProperty('--panel-bg', `rgba(${r},${g},${b},${ws.opacity})`)
       document.documentElement.style.setProperty('--text-primary', ws.fontColor)
@@ -84,61 +70,9 @@ export function SettingsPopover({ onClose }: SettingsPopoverProps) {
     updateWindowState({ alwaysOnTop: newLevel })
   }
 
-  const handleShortcutRecord = (e: React.KeyboardEvent) => {
-    e.preventDefault()
-    e.stopPropagation()
-    const parts: string[] = []
-    if (e.metaKey) parts.push('Command')
-    if (e.ctrlKey) parts.push('Control')
-    if (e.altKey) parts.push('Alt')
-    if (e.shiftKey) parts.push('Shift')
-    const key = e.key
-    if (!['Meta', 'Control', 'Alt', 'Shift'].includes(key)) {
-      parts.push(key === ' ' ? 'Space' : key.length === 1 ? key.toUpperCase() : key)
-      const combo = parts.join('+')
-      setIsRecording(false)
-      setShortcutError('')
-      api.windowSetShortcut(combo).then((res: any) => {
-        if (res?.error) {
-          setShortcutError(res.error)
-        } else {
-          setShortcut(combo)
-        }
-      })
-    }
-  }
-
   const handleLeadTimeChange = (minutes: number) => {
     setReminderLeadTime(minutes)
     api.configSet({ reminderLeadTime: minutes } as any)
-  }
-
-  const handleTargetChange = (target: 'command' | 'notes' | 'task') => {
-    setShortcutTarget(target)
-    api.configSet({ shortcutTarget: target } as any)
-  }
-
-  const handleTargetOptionChange = (option: 'last-edit' | 'new' | 'first') => {
-    setShortcutTargetOption(option)
-    api.configSet({ shortcutTargetOption: option } as any)
-  }
-
-  const checkForUpdates = async () => {
-    setUpdateStatus('checking')
-    try {
-      const res = await fetch('https://api.github.com/repos/KasparChen/meeemo/releases/latest')
-      if (!res.ok) { setUpdateStatus('latest'); return }
-      const data = await res.json()
-      const remote = (data.tag_name || '').replace(/^v/, '')
-      if (remote && remote !== version) {
-        setLatestVersion(remote)
-        setUpdateStatus('available')
-      } else {
-        setUpdateStatus('latest')
-      }
-    } catch {
-      setUpdateStatus('latest')
-    }
   }
 
   const THEME_DEFAULTS = {
@@ -226,57 +160,6 @@ export function SettingsPopover({ onClose }: SettingsPopoverProps) {
         <option value="bottom">Always on Bottom</option>
       </select>
 
-      {/* Shortcut */}
-      <div className="text-[10px] text-[var(--text-secondary)] font-semibold tracking-wider mt-3 mb-2">SHORTCUT</div>
-      {isRecording ? (
-        <div
-          className="w-full px-2 py-1.5 rounded text-sm text-center border-2 border-[var(--accent)] text-[var(--accent)] outline-none animate-pulse"
-          tabIndex={0}
-          ref={(el) => el?.focus()}
-          onKeyDown={handleShortcutRecord}
-          onBlur={() => setIsRecording(false)}
-        >
-          Press shortcut...
-        </div>
-      ) : (
-        <button
-          onClick={() => { setIsRecording(true); setShortcutError('') }}
-          className="w-full px-2 py-1.5 rounded text-sm bg-black/5 text-[var(--text-primary)] border border-[var(--border-color)] text-center cursor-pointer hover:bg-black/8"
-        >
-          {shortcut}
-        </button>
-      )}
-      {shortcutError && (
-        <div className="text-[10px] text-red-500 mt-1">{shortcutError}</div>
-      )}
-
-      {/* Shortcut Target */}
-      <div className="text-[10px] text-[var(--text-secondary)] font-semibold tracking-wider mt-3 mb-2">SHORTCUT TARGET</div>
-      <select
-        value={shortcutTarget}
-        onChange={(e) => handleTargetChange(e.target.value as any)}
-        className="w-full px-2 py-1.5 rounded text-sm bg-black/5 text-[var(--text-primary)] border border-[var(--border-color)] outline-none cursor-pointer"
-      >
-        <option value="command">Command Palette</option>
-        <option value="notes">Notes</option>
-        <option value="task">Tasks</option>
-      </select>
-
-      {shortcutTarget !== 'command' && (
-        <>
-          <div className="text-[10px] text-[var(--text-secondary)] font-semibold tracking-wider mt-2 mb-1">OPEN BEHAVIOR</div>
-          <select
-            value={shortcutTargetOption}
-            onChange={(e) => handleTargetOptionChange(e.target.value as any)}
-            className="w-full px-2 py-1.5 rounded text-sm bg-black/5 text-[var(--text-primary)] border border-[var(--border-color)] outline-none cursor-pointer"
-          >
-            <option value="last-edit">Last Edited</option>
-            <option value="new">New</option>
-            <option value="first">First</option>
-          </select>
-        </>
-      )}
-
       {/* Reminder Lead Time */}
       <div className="text-[10px] text-[var(--text-secondary)] font-semibold tracking-wider mt-3 mb-2">REMINDER</div>
       <select
@@ -293,62 +176,6 @@ export function SettingsPopover({ onClose }: SettingsPopoverProps) {
         <option value={120}>2 hours before</option>
         <option value={1440}>1 day before</option>
       </select>
-
-      {/* Storage path */}
-      <div className="border-t border-[var(--border-color)] mt-3 pt-2">
-        <div className="text-[10px] text-[var(--text-secondary)] font-semibold tracking-wider mb-2">STORAGE</div>
-        <div className="text-[10px] text-[var(--text-secondary)] truncate mb-1" title={storagePath}>
-          {storagePath}
-        </div>
-        <div className="flex gap-1">
-          <button
-            onClick={() => api.openStorage()}
-            className="flex-1 text-[10px] px-2 py-1 rounded bg-black/5 text-[var(--text-primary)] border border-[var(--border-color)] hover:bg-black/8 cursor-pointer"
-          >
-            Open Folder
-          </button>
-          <button
-            onClick={async () => {
-              const newPath = await api.changeStorage()
-              if (newPath) setStoragePath(newPath)
-            }}
-            className="flex-1 text-[10px] px-2 py-1 rounded bg-black/5 text-[var(--text-primary)] border border-[var(--border-color)] hover:bg-black/8 cursor-pointer"
-          >
-            Change...
-          </button>
-        </div>
-      </div>
-
-      {/* Version & Update */}
-      <div className="border-t border-[var(--border-color)] mt-3 pt-2">
-        <div className="flex items-center justify-between">
-          <span className="text-[10px] text-[var(--text-secondary)]">v{version}</span>
-          {updateStatus === 'idle' && (
-            <button
-              onClick={checkForUpdates}
-              className="text-[10px] text-[var(--accent)] hover:underline cursor-pointer"
-              style={{ border: 'none', background: 'none' }}
-            >
-              Check for updates
-            </button>
-          )}
-          {updateStatus === 'checking' && (
-            <span className="text-[10px] text-[var(--text-secondary)]">Checking...</span>
-          )}
-          {updateStatus === 'latest' && (
-            <span className="text-[10px] text-green-600">Up to date</span>
-          )}
-          {updateStatus === 'available' && (
-            <button
-              onClick={() => api.openUrl('https://github.com/KasparChen/meeemo/releases/latest')}
-              className="text-[10px] text-[var(--accent)] hover:underline cursor-pointer"
-              style={{ border: 'none', background: 'none' }}
-            >
-              v{latestVersion} available →
-            </button>
-          )}
-        </div>
-      </div>
     </div>
   )
 }
